Created splash fragment and changed the directories in the project (now they fit the basic MVVM pattern). Created navigation graph, added ktor client dependency. Added already created fragments to the navigation graph
