+++ .idea/misc.xml
@@ -12,6 +12,7 @@
         <entry key="..\:/Users/-/AndroidStudioProjects/FitnessMobileProject/app/src/main/res/layout/fragment_login.xml" value="0.33" />
         <entry key="..\:/Users/-/AndroidStudioProjects/FitnessMobileProject/app/src/main/res/layout/fragment_register.xml" value="0.2703125" />
         <entry key="..\:/Users/-/AndroidStudioProjects/FitnessMobileProject/app/src/main/res/layout/fragment_register_parameters.xml" value="0.25" />
+        <entry key="..\:/Users/-/AndroidStudioProjects/FitnessMobileProject/app/src/main/res/layout/fragment_splash.xml" value="0.35260416666666666" />
         <entry key="..\:/Users/-/AndroidStudioProjects/FitnessMobileProject/app/src/main/res/layout/item_program.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/-/AndroidStudioProjects/FitnessMobileProject/app/src/main/res/layout/register_toolbar.xml" value="0.36666666666666664" />
       </map>
