+++ app/src/main/java/application/workout/fitnessmobileproject/views/MainActivity.kt
@@ -1,8 +1,8 @@
-package application.workout.fitnessmobileproject
+package application.workout.fitnessmobileproject.views
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.lifecycle.ViewModel
+import application.workout.fitnessmobileproject.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
