+++ app/src/main/java/application/workout/fitnessmobileproject/views/fragments/HomeFragment.kt
@@ -1,4 +1,4 @@
-package application.workout.fitnessmobileproject.fragments
+package application.workout.fitnessmobileproject.views.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,17 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import application.workout.fitnessmobileproject.R
 
-class RegisterFragment : Fragment() {
-
+class HomeFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
     }
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_register, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
     }
 }
