+++ app/src/main/java/application/workout/fitnessmobileproject/views/fragments/RegisterFragment.kt
@@ -1,4 +1,4 @@
-package application.workout.fitnessmobileproject.fragments
+package application.workout.fitnessmobileproject.views.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import application.workout.fitnessmobileproject.R
-class RegisterParametersFragment : Fragment() {
+
+class RegisterFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -17,7 +18,6 @@ class RegisterParametersFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_register_parameters, container, false)
+        return inflater.inflate(R.layout.fragment_register, container, false)
     }
-
 }
