+++ app/src/main/java/application/workout/fitnessmobileproject/views/fragments/RegisterParametersFragment.kt
@@ -1,0 +1,23 @@
+package application.workout.fitnessmobileproject.views.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import application.workout.fitnessmobileproject.R
+class RegisterParametersFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_register_parameters, container, false)
+    }
+
+}
