+++ app/src/main/java/application/workout/fitnessmobileproject/views/fragments/SplashFragment.kt
@@ -1,4 +1,4 @@
-package application.workout.fitnessmobileproject.fragments
+package application.workout.fitnessmobileproject.views.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,8 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import application.workout.fitnessmobileproject.R
 
-class LoginFragment : Fragment() {
-
+class SplashFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
     }
@@ -18,6 +17,6 @@ class LoginFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login, container, false)
+        return inflater.inflate(R.layout.fragment_splash, container, false)
     }
 }
