+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.HomeFragment">
+    tools:context=".views.fragments.HomeFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/top_image"
