+++ app/src/main/res/layout/fragment_register.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.RegisterFragment">
+    tools:context=".views.fragments.RegisterFragment">
 
     <include
         android:id="@+id/toolbar"
