+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/splashFragment">
+
+    <fragment
+        android:id="@+id/splashFragment"
+        android:name="application.workout.fitnessmobileproject.views.fragments.SplashFragment"
+        android:label="fragment_splash"
+        tools:layout="@layout/fragment_splash" >
+        <action
+            android:id="@+id/action_splashFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
+        <action
+            android:id="@+id/action_splashFragment_to_homeFragment"
+            app:destination="@id/homeFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="application.workout.fitnessmobileproject.views.fragments.HomeFragment"
+        android:label="fragment_home"
+        tools:layout="@layout/fragment_home" />
+    <fragment
+        android:id="@+id/loginFragment"
+        android:name="application.workout.fitnessmobileproject.views.fragments.LoginFragment"
+        android:label="fragment_login"
+        tools:layout="@layout/fragment_login" >
+        <action
+            android:id="@+id/action_loginFragment_to_homeFragment"
+            app:destination="@id/homeFragment" />
+        <action
+            android:id="@+id/action_loginFragment_to_registerFragment"
+            app:destination="@id/registerFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/registerFragment"
+        android:name="application.workout.fitnessmobileproject.views.fragments.RegisterFragment"
+        android:label="fragment_register_parameters"
+        tools:layout="@layout/fragment_register_parameters" >
+        <action
+            android:id="@+id/action_registerFragment_to_registerParametersFragment"
+            app:destination="@id/registerParametersFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/registerParametersFragment"
+        android:name="application.workout.fitnessmobileproject.views.fragments.RegisterParametersFragment"
+        android:label="RegisterParametersFragment" >
+        <action
+            android:id="@+id/action_registerParametersFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
+    </fragment>
+</navigation>
